Move plugin manager outside of config folder
