+++ neovim/.config/nvim/init.lua
@@ -1,2 +1,2 @@
-require("config.plugin-manager") -- This needs to be at the top
+require("plugin-manager") -- This needs to be at the top
 require("config.options")
